backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
passlib[bcrypt]
python-jose[cryptography]
Pillow
aiofiles
python-dotenv
